+++ src/js/theme.js
@@ -47,8 +47,6 @@ class Util {
 }
 
 class FixIt {
-  #encrypted = typeof FixItDecryptor !== 'undefined';
-
   constructor() {
     this.config = window.config;
     this.data = this.config.data;
@@ -881,7 +879,7 @@ class FixIt {
   }
 
   initWatermark() {
-    this.config?.watermark?.enable &&
+    this.config.watermark?.enable &&
       new Watermark({
         content: this.config.watermark.content || '<img class="fixit-icon" src="/images/fixit.svg" alt="FixIt logo" /> FixIt Theme',
         appendTo: this.config.watermark.appendto || '.wrapper>main',
@@ -929,7 +927,13 @@ class FixIt {
         });
       }
     });
-    this.decryptor.init();
+    if (this.config.encryption?.shortcode) {
+      this.decryptor.addEventListener('decrypted', () => {
+        this.decryptor.initShortcodes();
+      })
+      this.decryptor.initShortcodes();
+    }
+    this.config.encryption?.all && this.decryptor.init();
   }
 
   onScroll() {
@@ -1029,9 +1033,9 @@ class FixIt {
 
   init() {
     try {
-      if (this.#encrypted) {
+      if (this.config.encryption) {
         this.initFixItDecryptor();
-      } else {
+      } else if(!this.config.encryption?.all) {
         this.initTwemoji();
         this.initDetails();
         this.initLightGallery();
@@ -1056,7 +1060,7 @@ class FixIt {
 
       window.setTimeout(() => {
         this.initComment();
-        if (!this.#encrypted) {
+        if (!this.config.encryption?.all) {
           this.initToc();
           this.initTocListener();
         }