+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 > Snapshot: 
 
 - :tada: Feat: add content encryption of pages ([#123](https://github.com/Lruihao/FixIt/issues/123))
+- :tada: Feat: add `fixit-encryptor` shortcode ([#123](https://github.com/Lruihao/FixIt/issues/123))
 - :sparkles: Feat: add total word count feature in section and remove from footer ([#124](https://github.com/Lruihao/FixIt/issues/124))
 - :sparkles: Feat: enhance auto toc feature ([#104](https://github.com/Lruihao/FixIt/issues/104) [#136](https://github.com/Lruihao/FixIt/issues/136))
 - :sparkles: Feat(SEO): add options to make output baidu_urls.txt file ([#138](https://github.com/Lruihao/FixIt/issues/138))